requirements.txt
Flask==3.1.0
graphviz==0.20.3
Pillow==11.1.0
